+++ database/migrations/2020_02_19_183502_create_twostep_auth_table.php
@@ -26,7 +26,7 @@ class CreateTwostepAuthTable extends Migration
     {
         Schema::create('twostep_auth', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('userId');
+            $table->unsignedBigInteger('userId');
             $table->string('authCode')->nullable();
             $table->integer('authCount');
             $table->boolean('authStatus')->default(0);
